1_chat_models/1_chat_model_basic.py
from dotenv import load_dotenv
from langchain_openai import ChatOpenAI

load_dotenv(override=True)

model = ChatOpenAI(model="gpt-4o-mini")

result = model.invoke("What is 2+2?")

# print(result)

print(result.content)