pyproject.toml
[project]
name = "learning-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langchain-openrouter>=0.2.1",
    "python-dotenv>=1.2.2",
]
